pyproject.toml
[project]
name = "agentic-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gradio>=5.44.1",
    "ipykernel>=6.30.1",
    "ipython>=9.5.0",
    "openai>=1.106.1",
    "openai-agents>=0.4.2",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.11",
]
